% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_calcFC.R
\name{calcFC}
\alias{calcFC}
\title{Calculate a forecast from a single statistical model}
\usage{
calcFC(fit.obj = NULL, data = NULL, fc.yr = NULL, settings = NULL,
  tracing = FALSE)
}
\arguments{
\item{fit.obj}{A list, created by \code{\link{fitModel}}, with model fitting
details for each age class.}

\item{data}{A list. Equal to the element named \code{data} from output of
\code{\link{prepData}}.}

\item{fc.yr}{An integer of length one. The forecast year. If NULL (the default), then calculate as the year after the last run year in the data set.}

\item{settings}{A list. The model-specific list of settings used in \code{\link{fitModel}}.}

\item{tracing}{A Boolean. Default is FALSE}
}
\description{
Calculate a forecast from a single statistical model
}
