% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_Sub_boots_functions.R
\name{createBoots}
\alias{createBoots}
\title{Bootstrap and reformat raw data}
\usage{
createBoots(dat.prepped, boot.type = c("meboot", "stlboot"),
  boot.n = 1000, plot.diagnostics = FALSE)
}
\arguments{
\item{plot.diagnostics}{}
}
\description{
Bootstrap and reformat raw data
}
\examples{
\dontrun{
data.use <- read.csv("../inst/extdata/SampleFile_WithAge.csv")  # read in raw data
data.use <- read.csv("../inst/extdata/FinalSampleFile_WithoutAge.csv")  # read in raw data
dat.prepped <-  prepData(datafile = data.use,out.labels="v2")
bootdata.reformat <- createBoots(dat.prepped = dat.prepped,boot.type = "meboot" )
length(bootdata.reformat)
bootdata.reformat[[1]]
}
}
