% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_DoRetro.R
\name{doRetro}
\alias{doRetro}
\title{Run a retrospective evaluation.}
\usage{
doRetro(model = NULL, data = NULL, retro.settings = NULL,
  fit.settings = NULL, fc.settings = NULL, tracing = FALSE,
  out.type = c("short", "full"), interval.n = 500,
  interval.quants = TRUE, pt.fc.in = NULL)
}
\arguments{
\item{model}{A character vector of length one. It is the model name and must match one of the model names in the \code{estimation.functions} list object.}

\item{data}{A data frame. Equivalent to the "data" element in the list output from \code{\link{prepData}}.}

\item{retro.settings}{A list}

\item{fit.settings}{A list}

\item{fc.settings}{A list}

\item{tracing}{A Boolean (default is FALSE)}

\item{out.type}{A character vector of length one. Can have value "short" (default) or "full". See value for more information.}

\item{interval.n}{An integer. The sample size for generating the retro-based interval. Default is 500.}

\item{interval.quants}{A Boolean (default is TRUE). If TRUE the output includes quantiles of the retro-based interval, rather than the full sample.}

\item{pt.fc}{}
}
\value{
Key output (with out.type="short") from this function includes two summary tables with performance measures from the dynamic retrospective:  retro.pm.all: for each age class, use as many retrospective years as possible (i.e more for younger ages). retro.pm.bal: balanced output, using the same retrospective years for all age classes and the total. If out.type="full", then the output also includes the full model fit info for each retro year
}
\description{
Run a retrospective evaluation.
}
