% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_doSampleFromInt.R
\name{doSampleFromInt}
\alias{doSampleFromInt}
\title{Get a sample based on bounds in a FC object}
\usage{
doSampleFromInt(fc.obj, interval.n = 1000, interval.quants = FALSE)
}
\arguments{
\item{fc.obj}{A list. Output of the function \code{\link{calcFC}}.}

\item{interval.n}{A numeric value to set sample size.}

\item{interval.quants}{If TRUE, calculate quantiles (p10,p25,p50,p75,p90)}
}
\value{
A data frame with samples (rows) by age class (columns)
}
\description{
Get a sample based on bounds in a FC object
}
\details{
MF to insert
}
