% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_fitModel.R
\name{fitModel}
\alias{fitModel}
\title{General Model Fitting Functon}
\usage{
fitModel(model = c("Naive", "Mechanistic", "SibRegSimple",
  "SibRegKalman", "SibRegLogPower", "TimeSeriesArima",
  "TimeSeriesExpSmooth"), data = NULL, settings = NULL,
  tracing = FALSE)
}
\arguments{
\item{model}{A character vector of length one. The name of a model, it has to
match one of the model names in the "estimation.functions" list objects.
which is generated from the script R/Module_Sub_EstimationFunctions.R.}

\item{data}{Same as the data element from the list object generated by
prepData()}

\item{settings}{A list. Model-specific list of settings, where applicable.}

\item{tracing}{A Boolean. Default is FALSE.}
}
\value{
A list.
}
\description{
This function applies the estimation step for the various
  models. It sets up the data inputs (e.g. loop through age classes, ) and
  calls the estimation subroutine then stores the output in a list object.
  Model-specific sub-routines live in the file
  "Module_Sub_EstimationFunctions.R".
}
\details{
FOR NOW: JUST MAKING THIS WORK WITH THE BASIC SIBLING REGRESSION AND
  KALMAN FILTER SIBLING REGRESSION
}
