% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_Sub_boots_functions.R
\name{fitModelandcalcFC}
\alias{fitModelandcalcFC}
\title{Function to apply fitModel() then calcFC(), and return the pt.fc}
\usage{
fitModelandcalcFC(
  data = NULL,
  fitmodel.args = list(model = NULL, settings = NULL),
  calcfc.args = list(fc.yr = NULL, settings = NULL)
)
}
\arguments{
\item{data}{output from \code{\link{prepData}}}

\item{fitmodel.args}{model fitting specs as per \code{\link{fitModel}}}

\item{calcfc.args}{forecast specs as per \code{\link{calcFC}}}
}
\value{
point forecast.
}
\description{
Function to apply fitModel() then calcFC(), and return the pt.fc
}
