% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_Sub_utils.R, R/utils.R
\name{load_or_install}
\alias{load_or_install}
\title{Package load or install and load}
\usage{
load_or_install(package_names)

load_or_install(package_names)
}
\arguments{
\item{package_names}{Character string of length one or more. The package name.}

\item{package_names}{Character string of length one or more. The package name.}
}
\value{
Nothing returned.

Nothing returned.
}
\description{
For installing and/or loading R packages

For installing and/or loading R packages
}
