% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_multiFC.R
\name{multiFC}
\alias{multiFC}
\title{Process multiple models and produce forecasts with summaries}
\usage{
multiFC(data.file, settings.list, do.retro = FALSE, retro.min.yrs = 15,
  out.type = c("short", "full"), int.type = "None", int.n = 100,
  boot.type = "meboot", tracing = FALSE)
}
\arguments{
\item{data.file}{The data file read in with \code{read.csv}.}

\item{settings.list}{A list with list sub-elements. See details.}

\item{do.retro}{A Boolean (default is FALSE). Run a retrospective}

\item{retro.min.yrs}{An integer of length one. Default is 15.}

\item{out.type}{A character vector of length one. The choices are: "short" or
"full". See value for details.}

\item{int.type}{Either "None","Retrospective","Prediction", or "Bootstrap"}

\item{int.n}{number of samples used for the interval calculations}

\item{boot.type}{= "meboot" or "stlboot", only applies if int.type = "Bootstrap"}

\item{tracing}{A Boolean (default is FALSE)}
}
\value{
A list is produced. If argument out.type = "short" and do.retro = FALSE,
then only generate a summary table of point forecasts.
  If out.type = "short" and do.retro= TRUE, then generate 3 versions of
  retrospective summary and array of fitted performance measures. If out.type
  ="full" and do.retro= TRUE, then also store the model fits for each retro
  year. If out.type ="short" and = TRUE, then TBD. If out.type ="full"
  and do.boot= TRUE, then TBD
}
\description{
Process multiple models and produce forecasts with summaries
}
\details{
The settings.list argument is a list defining each forecast model in
  separate lists that can be uniqely named. Each model list must have the
  elements  "model.type" and "settings". The settings element is also a list.
  For example: \code{settings.list= list(Naive1 =
  list(model.type="Naive",settings=list(avg.yrs=1)), SibRegSimple =
  list(model.type="SibRegSimple",settings=NULL))}
}
