% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_Sub_boots_functions.R
\name{plotBootSeries}
\alias{plotBootSeries}
\title{Plot a Summary of Bootstrapped Series}
\usage{
plotBootSeries(
  dat.prepped,
  boot.type = c("meboot", "stlboot"),
  age.which = "all"
)
}
\arguments{
\item{dat.prepped}{the $data element from the output of \code{\link{prepData}}}

\item{boot.type}{one of "meboot" or "stlboot"}

\item{age.which}{one of "all", "total", or specific age class (e.g. "Age 4")}
}
\value{
one or more plots.
}
\description{
Plot a Summary of Bootstrapped Series
}
