% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_plotModelFit.R
\name{plotModelFit}
\alias{plotModelFit}
\title{Plot a model fit...}
\usage{
plotModelFit(
  fit.obj,
  options = list(plot.which = "all", age.which = "all", plot.add = FALSE),
  fc.add = NULL,
  tracing = FALSE
)
}
\arguments{
\item{fc.add}{}
}
\value{

}
\description{
Plot a model fit...
}
