% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_Sub_boots_functions.R
\name{reformatBootData}
\alias{reformatBootData}
\title{Reformat bootstrapped input data to match normal format}
\usage{
reformatBootData(boot.output, prepData.ouput)
}
\arguments{
\item{boot.output}{A list of lists. Ouput of \code{\link{bootSeries}} can be
converted to a list to be compatible.}

\item{prepData.ouput}{A list. Output of \code{\link{prepData}}.}
}
\value{
A list with length equal to number of bootstrap samples (equaling the
  column count in each matrix in \code{boot.output}). Each subelement is a
  list and maintains the structure of \code{prepData.output}.
}
\description{
Reformat bootstrapped input data to match normal format
}
\examples{
\dontrun{
data.use <- read.csv("SampleFile_WithAge.csv")  # read in raw data
dat.prepped <-  prepData(datafile = data.use,out.labels="v2")
boot.n <- 1000
boot.output <- lapply(dat.prepped$data, function(x){
#name of age column varies, so use index:
	series.use <- x[,3]
createBoots(series = series.use, boot.type = "meboot", boot.n = boot.n , plot.diagnostics = FALSE )
createBoots(series = series.use, boot.type = "stlboot", boot.n = 1000 , plot.diagnostics = FALSE )
})
bootdata.reformat <- reformatBootData(boot.output, prepData.ouput = dat.prepped)
length(bootdata.reformat)
str(bootdata.reformat[[1]])
}
}
