% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{usePackage}
\alias{usePackage}
\title{Package load or install and load}
\usage{
usePackage(p)
}
\arguments{
\item{p}{Character string of length one. The package name.}
}
\value{
Nothing returned.
}
\description{
For installing and/or loading R packages
}
