% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_Sub_boots_functions.R
\name{bootSeries}
\alias{bootSeries}
\title{Create bootstrap matrix from a single series}
\usage{
bootSeries(series, boot.type = c("meboot", "stlboot"), boot.n = 1000,
  plot.diagnostics = FALSE, plot.type = c("all", "sample"))
}
\arguments{
\item{plot.type}{}
}
\description{
Create bootstrap matrix from a single series
}
